+++ pytest_pyodide/decorator.py
@@ -1,4 +1,5 @@
 import ast
+import functools
 import pickle
 import sys
 from base64 import b64decode, b64encode
@@ -9,8 +10,10 @@ from typing import Any, Protocol
 
 import pytest
 
+from .copy_files_to_pyodide import copy_files_to_emscripten_fs
 from .hook import ORIGINAL_MODULE_ASTS, REWRITTEN_MODULE_ASTS
 from .pyodide import JsException
+from .runner import _BrowserBaseRunner
 from .utils import package_is_built as _package_is_built
 
 
@@ -115,7 +118,7 @@ def _decode(result: str, selenium: SeleniumType) -> Any:
         raise ModuleNotFoundError(
             f"There was a problem with unpickling the return value/exception from your pyodide environment. "
             f"This usually means the type of the return value does not exist in your host environment. "
-            f"The original message is: {exc}. "
+            f"The original message is: {exc}."
         ) from None
 
 
@@ -409,3 +412,34 @@ class run_in_pyodide:
         wrapper = _create_outer_test_function(self._run_test, self._node)
 
         return wrapper
+
+
+def copy_files_to_pyodide(file_list, install_wheels=True, recurse_directories=True):
+    """A decorator that copies files across to pyodide"""
+
+    def wrap(fn):
+        @functools.wraps(fn)
+        def wrapped_f(*args, **argv):
+            # get selenium from args
+            selenium = None
+            for a in args:
+                if isinstance(a, _BrowserBaseRunner):
+                    selenium = a
+            for a in argv.values():
+                if isinstance(a, _BrowserBaseRunner):
+                    selenium = a
+            if not selenium:
+                raise RuntimeError(
+                    "copy_files_to_pyodide needs a selenium argument to your test fixture"
+                )
+            copy_files_to_emscripten_fs(
+                file_list,
+                selenium,
+                install_wheels=install_wheels,
+                recurse_directories=recurse_directories,
+            )
+            return fn(*args, **argv)
+
+        return wrapped_f
+
+    return wrap