+++ pytest_pyodide/server.py
@@ -11,11 +11,15 @@ import tempfile
 
 
 @contextlib.contextmanager
-def spawn_web_server(dist_dir):
+def spawn_web_server(dist_dir, extra_headers=None, handler_cls=None):
+    if not extra_headers:
+        extra_headers = {}
     tmp_dir = tempfile.mkdtemp()
     log_path = pathlib.Path(tmp_dir) / "http-server.log"
     q: multiprocessing.Queue[str] = multiprocessing.Queue()
-    p = multiprocessing.Process(target=run_web_server, args=(q, log_path, dist_dir))
+    p = multiprocessing.Process(
+        target=run_web_server, args=(q, log_path, dist_dir, extra_headers, handler_cls)
+    )
 
     try:
         p.start()
@@ -33,7 +37,7 @@ def spawn_web_server(dist_dir):
         shutil.rmtree(tmp_dir)
 
 
-def run_web_server(q, log_filepath, dist_dir):
+def run_web_server(q, log_filepath, dist_dir, extra_headers, handler_cls):
     """Start the HTTP web server
 
     Parameters
@@ -50,19 +54,33 @@ def run_web_server(q, log_filepath, dist_dir):
     sys.stdout = log_fh
     sys.stderr = log_fh
 
-    class Handler(http.server.SimpleHTTPRequestHandler):
-        def log_message(self, format_, *args):
-            print(
-                "[%s] source: %s:%s - %s"
-                % (self.log_date_time_string(), *self.client_address, format_ % args)
-            )
+    if not handler_cls:
 
-        def end_headers(self):
-            # Enable Cross-Origin Resource Sharing (CORS)
-            self.send_header("Access-Control-Allow-Origin", "*")
-            super().end_headers()
+        class DefaultHandler(http.server.SimpleHTTPRequestHandler):
+            def log_message(self, format_, *args):
+                print(
+                    "[%s] source: %s:%s - %s"
+                    % (
+                        self.log_date_time_string(),
+                        *self.client_address,
+                        format_ % args,
+                    )
+                )
 
-    with socketserver.TCPServer(("", 0), Handler) as httpd:
+            def end_headers(self):
+                # Enable Cross-Origin Resource Sharing (CORS)
+                self.send_header("Access-Control-Allow-Origin", "*")
+                for k, v in extra_headers.items():
+                    self.send_header(k, v)
+                if len(extra_headers) > 0:
+                    joined_headers = ",".join(extra_headers.keys())
+                    # if you don't send this, CORS blocks custom headers in javascript
+                    self.send_header("Access-Control-Expose-Headers", joined_headers)
+                super().end_headers()
+
+        handler_cls = DefaultHandler
+
+    with socketserver.TCPServer(("", 0), handler_cls) as httpd:
         host, port = httpd.server_address
         print(f"Starting webserver at http://{host}:{port}")
         httpd.server_name = "test-server"  # type: ignore[attr-defined]