+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 ## [Unreleased]
 
+- Add auto-setting of python version and runner version based on pyodide version.
+  [#66](https://github.com/pyodide/pytest-pyodide/pull/66)
+
+- Add support for custom headers in the pytest web server code, by setting
+  the `extra_headers` parameter in the `spawn_web_server` function.
+  [#39](https://github.com/pyodide/pytest-pyodide/pull/39)
+
+- Breaking: removed STANDALONE_REFRESH env variable which was used to
+  override `selenium_standalone` fixture with `selenium_standalone_refresh`.
+  [#65](https://github.com/pyodide/pytest-pyodide/pull/65)
+
+- Added command line option `--run-in-pyodide`. This will run a set of normal pytest tests in pyodide using the
+  same testing architecture used for running dedicated pyodide tests.
+  [#62](https://github.com/pyodide/pytest-pyodide/pull/62)
+
+- Breaking: `--runtime` commandline flag now requires runtimes to be comma-separated.
+  [#76](https://github.com/pyodide/pytest-pyodide/pull/76)
+
+- Add support for a custom `SimpleHTTPRequestHandler` class in the pytest
+  webserver code, by passing the `handler_cls` parameter in the
+  `spawn_web_server` function.
+  [#47](https://github.com/pyodide/pytest-pyodide/pull/47)
+
 ## [0.23.2] - 2022-11-14
 
 - Fixes for Python 3.11: there are some bugs with `ast.fix_missing_locations` in